# doc/api.yaml
openapi: 3.0.3
info:
  title: Tour Management API
  description: API for managing tourist circuits
  version: 'latest'
servers:
  - url: http://localhost:8080/api
...
paths:
...
  /interest_point/{interestId}:
    get:
      tags:
        - Interest Points
      summary: Get a specific interest point by ID
      parameters:
        - name: interestId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: interest point details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InterestPoint'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
    put:
      tags:
        - Interest Points
      summary: Update a specific interest point by ID
      parameters:
        - name: interestId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InterestPoint'
      responses:
        '200':
          description: interest point updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InterestPoint'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
...
components:
  schemas:
...
    Location:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
        latitude:
          type: number
          format: float
        longitude:
          type: number
          format: float
...
    InterestPoint:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
        description:
          type: string
        cost:
          type: number
          format: float
        location:
          $ref: '#/components/schemas/Location'
    Exception:
      type: object
      properties:
        type:
          type: string
        message:
          type: string
    BadRequestException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: BadRequestException
        message: Bad request
    NotAuthorizedException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: NotAuthorizedException
        message: Not authorized
    ResourceNotFoundException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: ResourceNotFoundException
        message: Resource of type <T> identified by <I> not found
    TooManyRequestsException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: TooManyRequestsException
        message: Too many requests
    InternalServerException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: InternalServerException
        message: Unexpected error
  responses:
    400:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BadRequestException'
    403:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/NotAuthorizedException'
    404:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ResourceNotFoundException'
    429:
      description: Too Many Requests
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/TooManyRequestsException'
    500:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/InternalServerException'